requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
lightgbm==4.5.0
joblib==1.4.2
pyyaml==6.0.2
matplotlib==3.9.2
seaborn==0.13.2

mlflow==2.16.2

fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
prometheus-fastapi-instrumentator==7.0.0
docker==7.1.0
python-multipart==0.0.12

dvc==3.55.2
pathspec<1.0

